refactor : 파일 줄이기
