+++ src/app/meetings/_components/modal/MeetingModalForm.tsx
@@ -18,7 +18,6 @@ import { QUERY_KEYS } from "@/constans/queryKey";
 import { getMeetingTypes } from "@/api/client";
 
 export function MeetingModalForm({
-  meetingTypeOptions: defaultMeetingTypeOptions = [],
   values,
   errors,
   isImageUploading,
@@ -29,42 +28,11 @@ export function MeetingModalForm({
   showImageMeta = true,
 }: MeetingBasicInfoSectionProps) {
 
-
   const { data: meetingTypes = [] } = useQuery<MeetingType[]>({
     queryKey: QUERY_KEYS.meetings.meetingType,
     queryFn: getMeetingTypes,
     staleTime: 1000 * 60 * 5,
   });
-
-
-  const meetingTypeOptions = (() => {
-    if (!values.category) {
-      return defaultMeetingTypeOptions;
-    }
-
-    const hasCurrentType = defaultMeetingTypeOptions.some(
-      (option) => option.value === values.category,
-    );
-
-    if (hasCurrentType) {
-      return defaultMeetingTypeOptions;
-    }
-
-    console.log([
-      {
-        value: values.category,
-        label: values.category || "현재 모임 종류",
-      },
-      ...defaultMeetingTypeOptions,
-    ]);
-    return [
-      {
-        value: values.category,
-        label: values.category || "현재 모임 종류",
-      },
-      ...defaultMeetingTypeOptions,
-    ];
-  })();
 
   return (
     <div className="space-y-5">
@@ -154,6 +122,24 @@ export function MeetingModalForm({
         hintText={errors.link}
       />
 
+      <InputCommon
+        id="capacity"
+        type="text"
+        inputMode="numeric"
+        label="모임 정원"
+        isRequired
+        placeholder="숫자만 입력해주세요"
+        value={values.capacity}
+        onChange={(event) => {
+          onChange({ capacity: event.target.value.replace(/[^0-9]/g, "") });
+        }}
+        onClear={() => {
+          onChange({ capacity: "" });
+        }}
+        isDestructive={Boolean(errors.capacity)}
+        hintText={errors.capacity}
+      />
+
       <div className="space-y-2">
         <p className="text-[14px] font-medium text-gray-800">
           이미지
@@ -188,23 +174,7 @@ export function MeetingModalForm({
         ) : null}
       </div>
 
-      <InputCommon
-        id="capacity"
-        type="text"
-        inputMode="numeric"
-        label="모임 정원"
-        isRequired
-        placeholder="숫자만 입력해주세요"
-        value={values.capacity}
-        onChange={(event) => {
-          onChange({ capacity: event.target.value.replace(/[^0-9]/g, "") });
-        }}
-        onClear={() => {
-          onChange({ capacity: "" });
-        }}
-        isDestructive={Boolean(errors.capacity)}
-        hintText={errors.capacity}
-      />
+
     </div>
   );
 }
