+++ src/constans/queryKey.ts
@@ -29,7 +29,6 @@ export const QUERY_KEYS = {
       "detail",
       `${String(meetingId)}`,
     ],
-    recommendationCandidates: ["meetings", "recommendationCandidates"] as const,
     list: ["meetings", "list"] as const,
     listParams: (params: { type: string; sortBy: string; sortOrder: string }) =>
       ["meetings", "list", params] as const,
